refactor(users): rename users type file
