+++ src/trakt/users.ts
